Remove legacy UI component
